refactor: add legacy, del a starategy, modift a state. work commit.
