--- src/PxSitl/include/PxSitl/Utils/Utils.h
@@ -1,22 +1,0 @@
-#if !defined(PX_SITL_UTILS_H)
-#define PX_SITL_UTILS_H
-
-#include <opencv2/core/base.hpp>
-
-namespace Utils {
-
-class Vision {
-private:
-  Vision() {}
-
-public:
-  static void colorThreshold(cv::Mat &inImage, cv::Mat &outImage) {
-    
-
-
-  }
-};
-
-} // namespace Utils
-
-#endif // PX_SITL_UTILS_H
