+++ src/PxSitl/include/PxSitl/vision/Strategy.hpp
@@ -1,5 +1,5 @@
-#if !defined(_STRATEGY_H_)
-#define _STRATEGY_H_
+#if !defined(_VISION_STRATEGY_H_)
+#define _VISION_STRATEGY_H_
 
 #include <iostream>
 
@@ -8,12 +8,10 @@ private:
 public:
   Strategy() {}
   virtual ~Strategy() {
-    std::cout << "Delete base strategy\n";
+    std::cout << "Delete strategy base\n";
   }
-  
-  void test() {}
 
   virtual void execute() = 0;
 };
 
-#endif // _STRATEGY_H_
+#endif // _VISION_STRATEGY_H_
