+++ src/PxSitl/include/PxSitl/vision/StrategyWait.hpp
@@ -2,20 +2,20 @@
 #define _STRATEGY_WAIT_H_
 
 #include "Strategy.hpp"
-#include "Utils/Utils.hpp"
 #include "ros/ros.h"
+#include "utils/Utils.hpp"
 
-class BallTrackingRos;
+// class BallTrackingRos;
 
-class StrategyWait: public Strategy
-{
+class StrategyWait : public Strategy {
 private:
   ros::Time _timer;
   ros::Duration _timeOut = ros::Duration(5);
-  BallTrackingRos *_context = nullptr;
+  // BallTrackingRos *_context = nullptr;
+  std::string _confFile = "/workspaces/Puzzle/src/PxSitl/data/config.yaml";
 
 public:
-  StrategyWait(BallTrackingRos *context):_context(context) {
+  StrategyWait() {
     _timer = ros::Time::now();
   }
