+++ src/PxSitl/src/vision/VideoHandler.cpp
@@ -1,4 +1,4 @@
-#include "../../include/PxSitl/Vision/VideoHandler.hpp"
+#include "../../include/PxSitl/vision/VideoHandler.hpp"
 
 VideoHandler& operator>>(VideoHandler &vh, cv::Mat &img) {
   vh.readColor(img);
