+++ src/puzzle_gazebo/launch/ball_tracking.launch
@@ -2,7 +2,7 @@
 <launch>
   <arg name="rviz_cong" value="$(find puzzle_gazebo)/rviz/ballTest.rviz"/>
 
-  <node name="ball_tracking" pkg="PxSitl" type="BallTrackingNode" output="log" required="true">
+  <node name="ball_tracking" pkg="PxSitl" type="BallTrackingNode" output="screen" required="true">
     <param name="color_topic" value="/camera/color/image_raw"/>
     <param name="depth_topic" value="/camera/aligned_depth_to_color/image_raw"/>
     <param name="filter_gain" value="0.0"/>
